# backend/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Backend
host: localhost:4000
...
basePath: /api
schemes:
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
     For accessing the API a valid JWT token must be passed in all the queries in
     the 'Authorization' header.
  
  
     A valid JWT token is generated by the API and retourned as answer of a call
     to the route /login giving a valid user & password.
  
  
     The following syntax must be used in the 'Authorization' header :
  
         Bearer xxxxxx.yyyyyyy.zzzzzz
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /procedure/{idProcedure}/polls:
    x-swagger-router-controller: polls.controller
    parameters:
      - name: idProcedure
        in: path
        description: "ID of the Procedure"
        required: true
        type: string
    post:
      tags:
        - Polls
      security:
        - Bearer: []
      x-security-scopes:
        - admin
        - user
        - dev
        - fede
      operationId: pollsControllerPost
      summary: "Create Poll"
      description: "Create a new Poll"
      parameters:
        - name: poll
          in: body
          required: true
          schema:
            $ref: '#/definitions/Poll'
      responses:
        "200":
          description: "Success"
        "403":
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - Polls
      operationId: pollsControllerGet
      summary: "Get polls"
      description: "Get all the polls"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: '#/definitions/Poll'
        "403":
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'
  /procedure/{idProcedure}/polls/{idPoll}:
    x-swagger-router-controller: polls.controller
    parameters:
      - name: idPoll
        in: path
        description: "ID of the poll"
        required: true
        type: string
      - name: idProcedure
        in: path
        description: "ID of the Procedure"
        required: true
        type: string
    get:
      tags:
        - Polls
      operationId: pollsControllerGetId
      summary: "Get poll"
      description: "Get only one polls"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: '#/definitions/Poll'
        "403":
          description: "Access Denied"
          schema:
            $ref: '#/definitions/Error'

    put:
      tags:
        - Polls
      summary: "Update a poll"
      operationId: pollsControllerPut
      description: "Update only one poll"
      security:
        - Bearer: []
      x-security-scopes:
        - admin
        - user
        - dev
        - fede
      parameters:
        - name: poll
          in: body
          required: true
          schema:
            $ref: '#/definitions/Poll'
      responses:
        200:
          description: "Success"
          
    delete:
      tags:
        - Polls
      summary: "Delete a poll"
      operationId: pollsControllerDelete
      description: "Delete only one poll"
      responses:
        200:
          description: "Success"
...
definitions:
...
  Poll:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
        example: "Womac Score"
      type:
        type: string
        example: "Score"
      detail:
        type: string
        example: "Pre Quirurgico"
      group:
        type: array
        items:
          $ref: '#/definitions/PollGroup'
...
  PollGroup:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      description:
      
        type: string
     
      subGroup:
        type: array
        items:
          $ref: '#/definitions/PollSubGroup'
  PollSubGroup:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      description:
        type: string
        example: "¿Qué grado de dolor tienes?"
      questions:
        type: array
        items:
          $ref: '#/definitions/Question'
      typeOfAnswer:
        $ref: '#/definitions/typeOfAnswer'
  typeOfAnswer:
    type: object
    properties:
      idTypeOfAnswer:
        type: string
      typeOfAnswer:
        type: string
        example: "TIPO DE REVISION, TIEMPO DE REVISION"
      description:
        type: string
        example: "En que instancia se encuentra para dicha revision."
      sampleValues:
        type: string
        example: "Ninguno, Leve, Moderado, Severo, Extremo"
      answers:
        type: array
        items:
          $ref: '#/definitions/answer'
      addable:
        type: boolean
        example: true
        type: string
  answer:
    type: object
    properties:
      id:
        type: string
      value:
        type: string
        example: "MODERADO"
      score:
        type: integer
        example: 2
  Question:
    type: object
    properties:
      id:
        type: string
      question:
        type: string
        example: "Subiendo y bajando escaleras"
...
  Error:
    type: object
    properties:
      message:
        type: string